requirements.txt
crispy-bootstrap5==2023.10
cryptography==42.0.5
Django==5.0
django-crispy-forms==2.1
django-recaptcha==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1